# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Documentation Api",
        "version": "1.0"
    },
    "paths": {
        "/api": {},
        "/api/register": {
            "post": {
                "summary": "Register User",
                "operationId": "8a56853624e025573120a09a4c75d468",
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "name": {
                                        "type": "string"
                                    },
                                    "email": {
                                        "type": "string"
                                    },
                                    "password": {
                                        "type": "string"
                                    },
                                    "confirm_password": {
                                        "type": "string"
                                    }
                                },
                                "type": "object",
                                "example": {
                                    "name": "atho",
                                    "email": "atho@gmail.com",
                                    "password": "Atho123",
                                    "confirm_password": "Atho123"
                                }
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Register successfully",
                        "content": {
                            "application/json": {
                                "schema": {},
                                "examples": {
                                    "result": {
                                        "summary": "An result object",
                                        "value": {
                                            "token": "371617ehwudhue37e673y"
                                        }
                                    }
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Invalid data",
                        "content": {
                            "application/json": {
                                "schema": {}
                            }
                        }
                    }
                }
            }
        },
        "/api/login": {
            "post": {
                "summary": "Login User",
                "operationId": "a3b306d14572d1f4bd6c064b3233e7b8",
                "requestBody": {
                    "content": {
                        "aplication/json": {
                            "schema": {
                                "properties": {
                                    "email": {
                                        "type": "string"
                                    },
                                    "password": {
                                        "type": "string"
                                    }
                                },
                                "type": "object",
                                "example": {
                                    "email": "atho@gmail.com",
                                    "password": "Atho123"
                                }
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Login successfully",
                        "content": {
                            "application/json": {
                                "schema": {},
                                "examples": {
                                    "result": {
                                        "summary": "An result object",
                                        "value": {
                                            "token": "371617ehwudhue37e673y"
                                        }
                                    }
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Invalid data",
                        "content": {
                            "application/json": {
                                "schema": {}
                            }
                        }
                    }
                }
            }
        },
        "/api/logout": {
            "delete": {
                "operationId": "fd694301e769d83fdd749695a3cc435b",
                "responses": {
                    "200": {
                        "description": "Success logout",
                        "content": {
                            "application/json": {
                                "schema": {},
                                "examples": {
                                    "result": {
                                        "summary": "An result object",
                                        "value": {
                                            "message": "Success logout"
                                        }
                                    }
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
        "/api/candidate": {
            "get": {
                "summary": "Get candidate",
                "operationId": "c792aa847309740c7570386e297d6ab7",
                "responses": {
                    "200": {
                        "description": "Login successfully",
                        "content": {
                            "application/json": {
                                "schema": {},
                                "examples": {
                                    "result": {
                                        "summary": "An result object",
                                        "value": {
                                            "name": "atho",
                                            "job_id": {
                                                "name": "fullstack"
                                            },
                                            "email": "atho@gmail.com",
                                            "phone": 89811378909
                                        }
                                    }
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            },
            "post": {
                "summary": "Create candidate",
                "operationId": "c4f871013efb0990f9053bc1b1b983f0",
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "name": {
                                        "type": "string"
                                    },
                                    "job_id": {
                                        "type": "integer"
                                    },
                                    "email": {
                                        "type": "string"
                                    },
                                    "phone": {
                                        "type": "integer"
                                    },
                                    "skill": {
                                        "type": "array",
                                        "items": {
                                            "type": "number"
                                        }
                                    }
                                },
                                "type": "object",
                                "example": {
                                    "name": "atho",
                                    "job_id": 2,
                                    "email": "atho@gmail.com",
                                    "phone": 84392493324,
                                    "skill": [
                                        1,
                                        2
                                    ]
                                }
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Create Candidate",
                        "content": {
                            "application/json": {
                                "schema": {},
                                "examples": {
                                    "result": {
                                        "summary": "An result object",
                                        "value": {
                                            "message": "Successfully create candidate"
                                        }
                                    }
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Invalid data",
                        "content": {
                            "application/json": {
                                "schema": {}
                            }
                        }
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
        "/api/candidate/{id}": {
            "get": {
                "summary": "Get candidate",
                "operationId": "97b55a7432cca8952d92db58165b3003",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "schema": {
                            "type": "number"
                        },
                        "examples": {
                            "int": {
                                "summary": "An int value",
                                "value": "1"
                            }
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Login successfully",
                        "content": {
                            "application/json": {
                                "schema": {},
                                "examples": {
                                    "result": {
                                        "summary": "An result object",
                                        "value": {
                                            "name": "atho",
                                            "job_id": {
                                                "name": "fullstack"
                                            },
                                            "email": "atho@gmail.com",
                                            "phone": 89811378909
                                        }
                                    }
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        }
    },
    "components": {
        "securitySchemes": {
            "sanctum": {
                "type": "http",
                "description": "Enter token in format (Bearer <token>)",
                "scheme": "Bearer",
                "name": "Authorization",
                "in": "header"
            }
        }
    }
}